requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.7
Django==4.1.2
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
Pillow==9.2.0
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
